refactor: modularize folder explorer
